pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GitHubNotificator"
version = "0.0.1"
authors = [
    {name = "Marcin Wiorko"},
]

dependencies = [
    "PyGithub >= 1.59.1",
    "pytest >= 7.4.0"
]

[project.scripts]
notify = "github_notificator.__main__:main"

[tool.ruff]
line-length = 120
fixable = ["ALL"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
#mypy_path = "."
files = "github_notificator, tests"
explicit_package_bases = true

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]


omit = [
    "__main__.py",
    "mutmut_config.py"
]
